remove inline prerender option
