+++ packages/next-server/types/index.d.ts
@@ -10,5 +10,5 @@ export type PageConfig = {
      */
     bodyParser?: { sizeLimit?: number | string } | false
   }
-  experimentalPrerender?: boolean | 'inline' | 'legacy'
+  experimentalPrerender?: boolean | 'legacy'
 }
