+++ test/integration/prerender/next.config.js
@@ -1,5 +1,3 @@
 module.exports = {
-  experimental: {
-    autoExport: true
-  }
+  target: 'serverless'
 }
