--- test/integration/prerender/pages/another.js
@@ -1,7 +1,0 @@
-const Page = ({ name }) => <p>Prerender page {name}</p>
-
-Page.getInitialProps = async () => ({ name: 'John Deux' })
-
-export const config = { experimentalPrerender: 'inline' }
-
-export default Page
