+++ test/integration/prerender/pages/another/index.js
