+++ test/integration/prerender/pages/blog/[post]/index.js
