--- test/integration/prerender/pages/nested/hello.js
@@ -1,7 +1,0 @@
-export const config = { experimentalPrerender: 'inline' }
-
-const Page = ({ title }) => <p>{title}</p>
-
-Page.getInitialProps = async () => ({ title: 'something' })
-
-export default Page
