--- test/integration/prerender/pages/nested/old-school.js
@@ -1,5 +1,0 @@
-const Page = () => <p>I'm just an old SSR page</p>
-
-Page.getInitialProps = () => ({})
-
-export default Page
