--- test/integration/spr-prerender/next.config.js
@@ -1,3 +1,0 @@
-module.exports = {
-  target: 'serverless'
-}
